Update constitution for voice agent, refine agent pipeline, remove completed specs
- Update constitution.md with voice agent architecture, models, rules, and directory structure
- Refine voice agent: improve fast_tts adapter, agent lifecycle, LLM brain, conversation manager
- Add gpu_manager.py for voice agent GPU memory management
- Remove completed voice agent task specs (TASK-VA-001 through TASK-VA-020)
- Remove completed functional/technical phase1 specs
- Remove cosyvoice SFT scripts and auto_optimize_cosyvoice benchmark

+++ src/voiceagent/asr/streaming.py
@@ -24,7 +24,7 @@ class StreamingASR:
             self.model = faster_whisper.WhisperModel(
                 config.model_name,
                 device="cuda",
-                compute_type="float32",
+                compute_type="float16",
             )
         except Exception as e:
             raise ASRError(
